Backport superclasses.
Before v0.79, the dump did not expose superclasses, causing the
Superclass field to be empty for classes before this version.
Backporting fills in this field with the earliest known value for each
class. It assumes that no changes were made to superclasses during this
time, which would be undetectable anyways.